requirements.txt
psycopg2-binary
confluent-kafka
simplejson
requests
coverage